[ep2to2ep1] epoch2 : TEMPORARELY move runTest.cc to P1_Sigma_sm_epem_mupmum
This is because I modified it in the vectorization branch in epoch1.
I want to have both branches identical before I apply the vectorization PR.

Note that epoch1 are now IDENTICAL except for epoch_process_id.h
